+++ changelog.md
@@ -4,6 +4,8 @@
     * Normalized and denoised existing ones
     * Added trading, hurt, snoring and coughing
     * Added sounds for females
+* Reputation is now the sum of all hearts
+  * Reputation has been renamed to hearts
 * Children now grow up in 16 days instead of 8 (configurable)
 * Babies no longer greet you
 * Added rose gold dust recipe and therefore a way to obtain rose gold